+++ condition-order-domain/src/test/java/me/caosh/condition/domain/model/strategy/container/TestTimeStrategy.java
@@ -23,7 +23,7 @@ class TestTimeStrategy implements TimeDrivenStrategy {
     }
 
     @Override
-    public int getStrategyId() {
+    public long getStrategyId() {
         return strategyId;
     }
